pyproject.toml
[tool.poetry]
name = "crm-system"
version = "0.1.0"
description = ""
authors = ["LonerLykos <loner.lykos.project@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.1"
mysqlclient = "^2.2.7"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
structlog = "^25.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"